pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stemforge"
version = "0.1.0"
description = "AI-powered stem separation, MIDI extraction, and music generation desktop application"
readme = "README.md"
license = { text = "MIT" }
requires-python = "==3.11.*"
authors = [
    { name = "StemForge Contributors" },
]
keywords = [
    "audio",
    "stem-separation",
    "midi",
    "music-generation",
    "demucs",
    "basicpitch",
    "musicgen",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Sound/Audio :: Editors",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# ---------------------------------------------------------------------------
# Runtime dependencies — REAL, VERIFIED, EXECUTION-READY
# ---------------------------------------------------------------------------
dependencies = [
    # Core PyTorch stack not uv manageable. Install with pip.
    #"torch==2.9.1+cu128",
    #"torchaudio==2.9.1+cu128",
    #"torchvision==0.20.1+cu128",

    # GitHub-pinned ML components (renamed distributions)
    "demucs @ git+https://github.com/facebookresearch/demucs.git@v4.0.1",
    "basic-pitch @ git+https://github.com/spotify/basic-pitch.git@v0.4.0",


    #Audiocraft and deps (required at import time)
    "audiocraft",
    "av @ git+https://github.com/PyAV-Org/PyAV.git@main",
    "num2words==0.5.14",
    "spacy==3.8.11",
    "transformers==5.2.0",
    "sentencepiece==0.2.0",
    "tokenizers==0.22.2",
    "torchmetrics==1.8.2",
    "omegaconf==2.3.0",
    "hydra-core==1.3.2",
    "flashy",

    # Demucs deps
    "dora-search==0.1.12",
    "submitit==1.5.4",
    "openunmix==1.3.0",
    "julius==0.2.7",
    "lameenc==1.8.1",

    # BasicPitch deps
    "pretty_midi==0.2.11",
    "mir_eval==0.8.2",
    "setuptools==59.8.0",

    # Audio I/O + DSP
    "soundfile==0.13.1",
    "numpy==1.26.4",
    "scipy==1.17.0",
    "librosa==0.11.0",
    "soxr==1.0.0",
    "pydub==0.25.1",

    # MIDI I/O
    "mido==1.3.3",

    # Vocal MIDI pipeline
    "faster-whisper>=1.0.0",

    # GUI
    "gradio==6.6.0",
    "gradio_client==2.1.0",

    # Config (tomllib for Python 3.10)
    "tomli>=2.0.0; python_version < '3.11'",
]

# ---------------------------------------------------------------------------
# Dependency groups (PEP 735) — uv dev tooling
# ---------------------------------------------------------------------------
[dependency-groups]
dev = ["pip"]

# ---------------------------------------------------------------------------
# Optional dependency groups (PEP 508 extras)
# ---------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]

# ---------------------------------------------------------------------------
# Console scripts
# ---------------------------------------------------------------------------
[project.scripts]
stemforge = "gui.app:main"

#----------------------------------------------------------------------------
# Tool hints ----------------------------------------------------------------
[tool.uv]
environments = ["python_version == '3.11' and sys_platform == 'linux'"]
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
[tool.uv.sources]
audiocraft = { git = "https://github.com/tsondo/audiocraft.git", rev = "fedora-av-compat" }
flashy = { path = "vendor/flashy_src" }
#---------------------------------------------------------------------------
# Hatchling build configuration
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["gui", "pipelines", "models", "utils"]
include = ["config.py"]
[tool.hatch.metadata]
allow-direct-references = true

# ---------------------------------------------------------------------------
# Ruff – linter and formatter
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["config", "gui", "pipelines", "models", "utils"]

# ---------------------------------------------------------------------------
# Mypy – static type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
